@@ -1,16 +1,16 @@
 <template>
   <div id="app">
-    <Todolist/>
+    <todolist/>
   </div>
 </template>
 
 <script>
-import Todolist from './views/Todolist.vue'
+import todolist from './views/Todolist.vue'
 
 export default {
   name: 'App',
   components: {
-    Todolist
+    todolist
   }
 }
 </script>
